Refactor: Extract BackIcon from appbar
